# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rproxy-api control API",
    "version": "see GET /capabilities and the release",
    "description": "The contract is docs/API.md; this document follows it."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid",
              "tls_config",
              "unsupported",
              "not_found",
              "already_exists",
              "static",
              "reserved",
              "bind_failed",
              "resolve_failed",
              "unauthorized",
              "forbidden",
              "internal"
            ]
          }
        },
        "description": "Every error response. `code` is stable; `error` is for people.",
        "required": [
          "error",
          "code"
        ],
        "additionalProperties": false
      },
...
      "ConfigReload": {
        "type": "object",
        "properties": {
          "added": {
            "type": "integer"
          },
          "removed": {
            "type": "integer"
          },
          "changed": {
            "type": "integer"
          },
          "unchanged": {
            "type": "integer"
          },
          "failed": {
            "type": "integer",
            "description": "Of the added and changed rules: those registered as failed"
          },
          "restart_needed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rules": {
            "type": "integer"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          }
        }
      },
      "Finding": {
        "type": "object",
        "properties": {
          "rule": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ConfigReloadError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          }
        },
        "required": [
          "code",
          "error"
        ]
      },
      "AcmeCertStatus": {
        "type": "object",
        "description": "State of an ACME certificate.",
        "properties": {
          "resolver": {
            "type": "string"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "valid",
              "renewing",
              "error"
            ],
            "description": "pending: not issued yet (a self-signed stand-in is served); renewing: due for renewal; error: the last attempt failed (a certificate issued earlier stays in use)"
          },
          "not_after": {
            "type": "string",
            "format": "date-time"
          },
          "renew_at": {
            "type": "string",
            "format": "date-time"
          },
          "next_attempt": {
            "type": "string",
            "format": "date-time",
            "description": "After a failure, or while the rate limit holds the order back"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "resolver",
          "domains",
          "state"
        ]
      },
      "Acme": {
        "type": "object",
        "description": "global.acme: names and states only. Secrets (DNS API keys, EAB keys) and their files are never shown.",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "directory": {
                  "type": "string"
                },
                "contact": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "eab": {
                  "type": "boolean"
                },
                "allowed_names": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "registered": {
                  "type": "boolean"
                }
              }
            }
          },
          "dns_providers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "powerdns",
                    "http"
                  ]
                },
                "zones": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "allowed_names": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "resolvers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "account": {
                  "type": "string"
                },
                "challenge": {
                  "type": "string",
                  "enum": [
                    "http-01",
                    "tls-alpn-01",
                    "dns-01"
                  ]
                },
                "dns_provider": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "certificates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcmeCertStatus"
            }
          },
          "rate_limit": {
            "type": "object",
            "properties": {
              "orders": {
                "type": "integer"
              },
              "period_secs": {
                "type": "integer"
              },
              "used": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/config/reload": {
      "post": {
        "summary": "Apply the settings file RPROXY_CONFIG now and answer what happened (admin; by default only over the Unix socket, RPROXY_API_RELOAD_UNIX_ONLY)",
        "responses": {
          "200": {
            "description": "Applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReload"
                }
              }
            }
          },
          "400": {
            "description": "The settings have a mistake; nothing changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReloadError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the admin scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "No settings file is configured (code no_config)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/acme": {
      "get": {
        "summary": "ACME accounts, DNS providers, resolvers and certificates (rules:read); names and states only, never a secret",
        "responses": {
          "200": {
            "description": "global.acme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acme"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/acme/renew": {
      "post": {
        "summary": "Renew an ACME certificate now, within the rate limit (acme:write; by default only over the Unix socket)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolver": {
                    "type": "string"
                  },
                  "domains": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "resolver",
                  "domains"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The renewal is started in the background (see GET /acme)"
          },
          "400": {
            "description": "Unknown resolver, or names outside the allowlists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the acme:write scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No rule uses this certificate, or global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/acme/accounts/{name}/register": {
      "post": {
        "summary": "Create the account at the CA, or find the existing one for its key (acme:write; by default only over the Unix socket)",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "An account of global.acme.accounts"
          }
        ],
        "responses": {
          "200": {
            "description": "Done"
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the acme:write scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such account, or global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The CA refused or could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/acme/accounts/{name}/deactivate": {
      "post": {
        "summary": "Deactivate the account at the CA and set its key aside; the next order creates a new account (acme:write; by default only over the Unix socket)",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "An account of global.acme.accounts"
          }
        ],
        "responses": {
          "200": {
            "description": "Done"
          },
          "401": {
            "description": "Missing, unknown or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the acme:write scope, or the request came over TCP while RPROXY_API_RELOAD_UNIX_ONLY is true",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such account, or global.acme is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The CA refused or could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}